requirements.txt
-i https://pypi.org/simple
apistar==0.5.41
astroid==2.1.0
certifi==2018.11.29
chardet==3.0.4
click==7.0
colorlog==4.0.2
idna==2.8
isort==4.3.4
jinja2==2.10
lazy-object-proxy==1.3.1
markupsafe==1.1.0
mccabe==0.6.1
pylint==2.2.2
python-dotenv==0.10.1
pyyaml==3.13
requests==2.21.0
six==1.12.0
urllib3==1.24.1
werkzeug==0.14.1
whitenoise==4.1.2
wrapt==1.11.1
